Completed ch14 code (strings)
I added all functions and code related to this chapter. But the interpreter is not yet able to perform print, error and load correctly. I'm done with this for today. Tomorrow, when my head is not aching anymore, I'll hunt down all typos and stuff I probably missed while writing the new functions down.